Move fetching code to ServiceDataLayer
